"hide" mouse cursor
Mouse cursor may end up covering important areas of the screen (i.e. insta-monkey end screen in chimps).
In that case the bot will fail to match the image from /assets and it will hang up.
To fix this we just move the mouse to the bottom right corner every time we are done clicking something.